Typings Support!
I'm wanting to try upgrading things to be statically typed with `tsc`, it just needs a little setup to do so. It's a very great thing that it's already ESM-ready, that helps a bunch! The only challenging part is that some of the JSDoc definitions aren't TS compatible, so it's giving some unexpected results, even if they are really typed out. So having misleading types can be hard too, in comparison to only untyped code.

I was working on adding typings support to 'ConvertJavaTextureToBedrockApi' first, but that depends on this one, so I have to make this one type-safe first.